// apps/k-pay-app/src/types/graphql.ts
export enum SupportedProviders {
  MTN_MOMO = 'MTN_MOMO',
  M_PESA = 'M_PESA',
  ORANGE = 'ORANGE',
  AIRTEL = 'AIRTEL',
}

export interface WithdrawalInput {
  walletId: string;
  currencyCode: string;
  amount: number;
  provider: SupportedProviders;
  receiverPhone: string;
  description?: string;
  paymentPin?: string;
}